Parse 10 card correctly
Before, the 10 card would be parsed incorrectly. For instance, for a ten
of diamonds (10d), the first character (1) would be read as the pips,
then the second character (0) would be read as the suit.

This changes the parser to read the first till the second last character
as the pips (10), then the last character as the suit (d).